backend/requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Werkzeug==1.0.1
requests==2.22.0
flask-cors==3.0.10
Flask-SSE==1.0.0
python-dotenv==0.19.1
Pillow==7.0.0
opencv-python==4.5.4.58
image==1.5.33
pytesseract==0.3.7
numpy==1.21.3
pytest-cov==2.12.1
pytest==6.2.4
gevent==21.8.0
matplotlib==3.4.3
gunicorn==20.1.0
